final/sv_request/requirements.txt
python-dotenv
redis
celery
progressbar2
moviepy
Pillow==7.1.2